The commit message doesn't match the provided diff. Please consider the following corrected commit message:
Add Assignment 2 instructions

A new file, Assignment_2_instructions.md, has been created to detail the instructions for Assignment 2. This includes steps to be followed for creating zonal statistics, specific instructions on which data to include and exclude, and a question to verify understanding of the output unit. A list of crop types corresponding to specific codes is also provided for reference.