Adjust Jupyter settings and refactor assignment code
This commit modifies both the Jupyter settings and the content of an assignment. It adds a new Jupyter settings file, enhancing the connection with Jupyterhub and its configurations on the server. In addition, the 'Assignment2_zonal.ipynb' notebook is refactored, adding more details, updating execution count and time, and improving its overall structure to increase clarity and readability. Lastly, the Python packages required for the project are updated in 'wb_installs.py'.