Change buffer distance and update comments in notebooks
Buffer distance input in `create_buffer` function call was changed from '10560ft' to 2 in both `jason_eda.ipynb` and `useful_code_commands.ipynb`. In `useful_code_commands.ipynb`, additional commenting was also provided, giving an option to switch between the different buffer distances. This ensures the use of the correct unit of measurement depending on the Coordinate System in use.