Remove BokehJS loading scripts from jason_eda_t1.ipynb
The diff reveals extensive BokehJS related scripts were removed from the code in the jason_eda_t1.ipynb file. This suggests that these scripts, which are responsible for loading BokehJS libraries and running scripts when BokehJS is finished loading, are no longer needed in this file. Removal of these scripts potentially simplifies the code and could either be due to a refactoring effort, or because the dependencies on BokehJS in this file have been completely removed.