Add new helper functions for data exploration and preprocessing
A new python file called 'windbreaks_helpers.py' was added and linked to the project. It includes functions for advanced inspection of pandas dataframes, geodataframes, data processing, and general outputs to facilitate understanding and data preprocessing. It includes functions to convert dataframes to shapefiles, create buffer geometries, and to filter data. This file was then imported into the 'jason_eda_preprocess.ipynb' file.