Add deprecated EDA preparation script
A new Exploratory Data Analysis (EDA) preparation script has been added. This script is considered deprecated, designed to configure and set up an environment suitable for data analysis process. It imports and instills necessary tools and libraries for this purpose. Please be aware that the use of this script is not recommended as it is outdated, and newer versions of the script may exist.